requirements.txt
streamlit
opencv-python
numpy
tensorflow
pillow
paho-mqtt
requests
urllib3